MYR-736: moved interface to its own folder and proper importing
