--- src/components-v2/button/ContainedButton-interfaces.ts
@@ -1,10 +1,0 @@
-export enum ButtonSize {
-  SMALL = 'small',
-  MEDIUM = 'medium',
-  LARGE = 'large',
-}
-
-export enum ButtonColor {
-  PRIMARY = 'primary',
-  SECONDARY = 'secondary',
-}
