+++ src/components-v2/button/ContainedButton.tsx
@@ -1,35 +1,10 @@
+import {action} from '@storybook/addon-actions';
+
 import React from 'react';
 
 import Button from '@material-ui/core/Button';
 
-import {ButtonSize, ButtonColor} from './ContainedButton-interfaces';
-
-interface ContainedButtonProps {
-  /**
-   * Define the base color of the button
-   */
-  color?: ButtonColor;
-
-  /**
-   * How large should the button be?
-   */
-  size?: ButtonSize;
-
-  /**
-   * Button accessibility label
-   */
-  ariaLabel?: string;
-
-  /**
-   * Button label
-   */
-  children: React.ReactNode;
-
-  /**
-   * Optional click handler
-   */
-  onClick?: () => void;
-}
+import {ButtonSize, ButtonColor, ButtonProps} from './button.interfaces';
 
 const ContainedButton = ({
   color = ButtonColor.PRIMARY,
@@ -37,9 +12,14 @@ const ContainedButton = ({
   ariaLabel = 'contained-button',
   children,
   ...props
-}: ContainedButtonProps): JSX.Element => {
+}: ButtonProps): JSX.Element => {
   return (
-    <Button color={color} size={size} aria-label={ariaLabel} variant="contained">
+    <Button
+      color={color}
+      size={size}
+      aria-label={ariaLabel}
+      variant="contained"
+      onClick={action('clicked')}>
       {children}
     </Button>
   );
